+++ CHANGELOG.md
@@ -17,6 +17,9 @@ Config:
 - [BC Break] Added `fr3d_ldap.user.usernameAttribute` for indicate the attribute which holds the username.
   Previously was the first entry in `attributes`
 
+Validator:
+- Removed unused `property` option
+
 ### v2.0.0
 
 Minimum Requirements: